Make for just extend detail
