+++ addon/components/detail.js
@@ -10,16 +10,35 @@ import {validate} from '../actions'
 
 import _ from 'lodash'
 import Ember from 'ember'
-const {A, Component, Logger} = Ember
+const {Component} = Ember
 import computed, {readOnly} from 'ember-computed-decorators'
 import PropTypeMixin, {PropTypes} from 'ember-prop-types'
 import dereference from '../dereference'
 import {getDefaultView} from '../generator'
-import validateView, {validateModel, validateValue} from '../validator/index'
+import validateView, {validateModel} from '../validator/index'
 import {deemberify, recursiveObjectCreate} from '../utils'
 
+/**
+ * Determine if an object is an Ember.Object or not
+ * @param {Object|Ember.Object} object - object to check
+ * @returns {Boolean} whether or not object is an Ember.Object
+ */
+function isEmberObject (object) {
+  return !_.isEmpty(object) && !_.isPlainObject(object)
+}
+
+const builtInRenderers = {
+  PropertyChooser: 'frost-bunsen-property-chooser'
+}
+
 export default Component.extend(PropTypeMixin, {
-  classNames: ['frost-bunsen-detail', 'inline'],
+  // ==========================================================================
+  // Dependencies
+  // ==========================================================================
+
+  // ==========================================================================
+  // Properties
+  // ==========================================================================
 
   propTypes: {
     model: PropTypes.oneOf([
@@ -43,91 +62,78 @@ export default Component.extend(PropTypeMixin, {
 
   getDefaultProps () {
     return {
+      classNames: ['frost-bunsen-detail'],
       renderers: {},
-      validators: []
+      validators: [],
+      value: null
     }
   },
 
-  /**
-   * Get the view (from props or from generator)
-   *
-   * @param {BunsenView} view - the view to use (if given)
-   * @param {BunsenModel} model - the model schema to use to generate a view (if view is undefined)
-   * @returns {BunsenView} the view
-   */
-  getView (view, model) {
-    if (!_.isEmpty(view)) {
-      return view
-    }
+  // ==========================================================================
+  // Computed Properties
+  // ==========================================================================
 
-    return getDefaultView(model)
+  @readOnly
+  @computed('renderers')
+  allRenderers (renderers) {
+    const passedInRenderers = renderers || {}
+    return _.assign({}, builtInRenderers, passedInRenderers)
+  },
+
+  @readOnly
+  @computed('errors')
+  renderErrors (errors) {
+    return errors || {}
+  },
+
+  @readOnly
+  @computed('model')
+  renderModel (model) {
+    return dereference(model || {}).schema
   },
 
   @readOnly
   @computed('model', 'view')
+  /**
+   * Get the view to render (generate one if consumer doesn't supply a view)
+   * @param {BunsenModel} model - the model schema to use to generate a view (if view is undefined)
+   * @param {BunsenView} view - the view to use (if given)
+   * @returns {BunsenView} the view to render
+   */
   renderView (model, view) {
-    return recursiveObjectCreate(
-      this.getView(view, model)
-    )
+    view = !_.isEmpty(view) ? view : getDefaultView(model)
+    return isEmberObject(view) ? view : recursiveObjectCreate(view)
   },
 
-  /**
-   * Validate the model given as attributes
-   */
-  validateProps () {
-    const model = this.get('model')
-    const renderers = this.get('state.renderers') || {}
-    const value = this.get('value')
-    const view = this.get('renderView')
-
-    let result = validateModel(model)
-
-    if (result.errors.length === 0) {
-      result = validateView(view, model, _.keys(renderers))
-    }
-
-    if (result.errors.length === 0 && value) {
-      result = validateValue(value, model, false)
-    }
-
-    if (_.isArray(result.errors) && result.errors.length) {
-      Logger.log('-BUNSEN ERRORS------------------------')
-      Logger.log('   `----.                ')
-      Logger.log('   :::::::.              ')
-      Logger.log(' `::O::::::-`            ')
-      Logger.log(' -:::::::::::.           ')
-      Logger.log(' `--:-. ::::::-`     ..  ')
-      Logger.log('       .::::::::-..`.::  ')
-      Logger.log('       .:::::::::::::-`  ')
-      Logger.log('       -:::::::::::-`    ')
-      Logger.log('        .:-``-::.        ')
-      Logger.log('       `..  `...         ')
-      for (let error of result.errors) {
-        Logger.warn(`${error.message} (${error.path})`)
-      }
-      Logger.log('--------------------------------------')
-    }
+  @readOnly
+  @computed('renderView')
+  cellConfig () {
+    return this.get('renderView.rootContainers.0')
   },
 
+  @readOnly
+  @computed('allRenderers', 'renderView')
   /**
-   * Ensure we are working with POJO's
+   * Get store
+   * @param {Object} renderers - renderer to component mapping
+   * @param {BunsenView} view - view being rendered
+   * @returns {Object} store
    */
-  fixPropTypes () {
-    ;[
-      'model',
-      'renderers',
-      'value',
-      'view'
-    ].forEach((key) => {
-      let object = this.get(key)
-
-      if (!_.isEmpty(object) && !_.isPlainObject(object)) {
-        object = deemberify(object)
-        this.set(key, object)
-      }
+  store (renderers, view) {
+    return Ember.Object.create({
+      formValue: this.get('reduxStore').getState().value,
+      renderers,
+      view
     })
   },
 
+  // ==========================================================================
+  // Functions
+  // ==========================================================================
+
+  /**
+   * Keep UI in sync with updates to redux store
+   */
   storeUpdated () {
     const onChange = this.get('onChange')
     const onValidation = this.get('onValidation')
@@ -136,7 +142,7 @@ export default Component.extend(PropTypeMixin, {
 
     this.setProperties({
       errors,
-      value
+      renderValue: value || {}
     })
 
     if (onChange) {
@@ -149,101 +155,64 @@ export default Component.extend(PropTypeMixin, {
   },
 
   /**
-   * Validate model and view when we first get them
+   * Setup redux store
    */
   init () {
     this._super()
 
-    this.fixPropTypes()
-
     const reduxStore = createStoreWithMiddleware(reducer)
 
-    this.setProperties({
-      errors: {},
-      reduxStore
-    })
-
+    this.set('reduxStore', reduxStore)
     reduxStore.subscribe(this.storeUpdated.bind(this))
-
-    const value = this.get('value') || {}
-
-    reduxStore.dispatch(
-      validate(null, value, this.get('renderModel'), this.get('validators'))
-    )
-
-    const passedInRenderers = this.get('renderers')
-    const renderers = _.assign({}, passedInRenderers)
-    this.set('state', Ember.Object.create({
-      propValidationResult: {
-        errors: A([]),
-        warnings: A([])
-      },
-      renderers
-    }))
-
-    this.fixPropTypes()
-    this.validateProps()
   },
 
   /**
-   * Validate model and view when we get updated ones
-   * @param {Object} attrs - A hash of the old attributes and the new attributes
+   * Validate properties
    */
-  didUpdateAttrs () {
-    this._super(...arguments)
-    // const passedInRenderers = this.get('renderers') || {}
-    // const renderers = _.assign({}, builtinRenderers, passedInRenderers)
+  validateProps () {
+    const model = this.get('model')
+    const modelPojo = isEmberObject(model) ? deemberify(model) : model
+    const renderers = this.get('allRenderers')
+    const view = this.get('renderView')
 
-    this.validateProps()
-    // if (!_.isEqual(this.get('state.renderers'), renderers)) {
-    //   this.set('state.renderers', renderers)
-    // }
+    let result = validateModel(modelPojo)
+
+    if (result.errors.length === 0) {
+      const viewPojo = isEmberObject(view) ? deemberify(view) : view
+      result = validateView(viewPojo, model, _.keys(renderers))
+    }
+
+    this.set('propValidationResult', result)
+  },
+
+  /**
+   * Keep value in sync with store and validate properties
+   */
+  didReceiveAttrs () {
+    this._super(...arguments)
 
     const reduxStore = this.get('reduxStore')
     const value = this.get('value')
+    const hasValueChanged = !_.isEqual(value, reduxStore.getState().value)
 
-    if (value === undefined) {
-      this.set('value', reduxStore.getState().value)
-      return
+    if (hasValueChanged) {
+      reduxStore.dispatch(
+        validate(null, value || {}, this.get('renderModel'), this.get('validators'))
+      )
     }
 
-    if (_.isEqual(value, reduxStore.getState().value)) {
-      return
-    }
-
-    reduxStore.dispatch(
-      validate(null, value, this.get('renderModel'), this.get('validators'))
-    )
+    this.validateProps()
   },
 
-  willUpdate () {
-    this.fixPropTypes()
-  },
+  // ==========================================================================
+  // Events
+  // ==========================================================================
 
-  @readOnly
-  @computed('model')
-  renderModel (model) {
-    return dereference(model || {}).schema
-  },
-
-  @readOnly
-  @computed('state.renderers', 'renderView')
-  store (renderers, view) {
-    return Ember.Object.create({
-      renderers,
-      view
-    })
-  },
-
-  @readOnly
-  @computed('renderView')
-  cellConfig () {
-    return this.get('renderView.rootContainers.0')
-  },
+  // ==========================================================================
+  // Actions
+  // ==========================================================================
 
   actions: {
-    onChange () {
-      //
-    }
+    onChange () {}
   }
 })
